Landing: remove farewell-solo route on Login request which is default
This creates lots of issues currently since Login is the default route
on most of the places and it ends up with the blog post. Since it's
already removed as a feature on production we don't need to worry about
this redirection, instead we can guide these users who has an issue with
solo logins in support channels.